refactor(worktree): lift the rename concern out of workspace_ops, under the CI size cap
`xtask file-size-lint` runs in CI (.github/workflows/ci.yml:130) with a 3000-LOC
hard cap, and `workspace_ops.rs` had been over it since b9d19ea9:

    FAIL  3176 LOC  apps/desktop/src/shell/workspace/workspace_ops.rs

Neither of the two phases that pushed it there noticed, because the gates being
run were `cargo test` and `cargo clippy` — which were green, and are not what CI
checks.

The rename flow moves to `rename_ops.rs` whole: the three pure helpers, the
seven `WorkspaceRoot` methods, and the seven unit tests. A concern rather than a
pile — the rename dialog, the inline double-click rename and the refusal's
label-only escape hatch are one flow, and `merge_ops.rs` will sit beside it.
`resolve_project_for_workspace` becomes `pub(crate)` so both can reach it.

Pure move: no behaviour changes, no signatures change. 2709 LOC, gate clean.
