+++ apps/desktop/src/shell/workspace/mod.rs
@@ -3,6 +3,7 @@
 
 pub mod add_project_dialog;
 pub mod project_picker;
+pub mod rename_ops;
 pub mod session_merge;
 pub mod workspace_dialog;
 pub mod workspace_ops;
